feat: writing
